requirements.txt
anyio==3.5.0
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
fastapi==0.71.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-dialogflow==2.10.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.12.0
httpcore==0.14.4
httptools==0.3.0
httpx==0.21.3
idna==3.3
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
watchgod==0.7
websockets==10.1
